third_party/WebKit/LayoutTests/fast/css-grid-layout/percent-of-indefinite-track-size-in-auto-expected.txt

CONSOLE WARNING: line 48: Percentages row tracks and gutters for indefinite height grid containers will be resolved against the intrinsic height instead of being treated as auto and zero respectively. This change will happen in M70, around October 2018. See https://www.chromestatus.com/feature/6708326821789696 for more details.
This test checks that percentages of indefinite sizes are treated as min-content (for the min track sizing function) or max-content (for the max track sizing function).

XXX
XXX
PASS
XXX
XX XX
PASS
XXX
XX XX
PASS
XXXX
X
PASS
XXXX
X
PASS
XXX
XX XX
XXXX
PASS
